Pass status check if cards removed from PR
 ## Summary
If we initially add Trello card links into a Pull Request but later
remove them, we're unable to remove the status check as GitHub doesn't
expose an endpoint for doing this. The best we can/should do in this
instance is, probably, flip the status check to 'passing' so that it no
longer appears to be blocking a merge.